+++ crates/lune-std/src/require/mod.rs
@@ -1,4 +1,5 @@
 mod loader;
 mod resolver;
 
+#[allow(unused_imports)]
 pub(crate) use self::resolver::RequireResolver;